Add Atom feeds for new beatmaps (#39)
Inserts a link in the head of every page to a feed of newly-uploaded
beatmaps. Additionally, when viewing a specific user's uploads, a
user-specific feed will be inserted.

Browsers/browser extensions will detect this and provide a button to
subscribe to the feeds.

The feed itself contains the beatmap image, information, as well as an
enclosure link to the actual beatmap zip file. With compatible feed
readers, this will enable easy auto-downloading.

Atom was chosen over RSS because it's a much cleaner specification and
pretty much every feed reader supports both anyway.

Fixes #20